+++ src/main/java/cart/dto/request/ProductRequest.java
@@ -1,5 +1,6 @@
 package cart.dto.request;
 
+import java.math.BigDecimal;
 import javax.validation.constraints.NotBlank;
 import javax.validation.constraints.NotNull;
 
@@ -9,7 +10,7 @@ public class ProductRequest {
     private String name;
 
     @NotNull
-    private Long price;
+    private BigDecimal price;
 
     @NotNull
     private String imageUrl;
@@ -17,7 +18,7 @@ public class ProductRequest {
     public ProductRequest() {
     }
 
-    public ProductRequest(String name, Long price, String imageUrl) {
+    public ProductRequest(String name, BigDecimal price, String imageUrl) {
         this.name = name;
         this.price = price;
         this.imageUrl = imageUrl;
@@ -27,7 +28,7 @@ public class ProductRequest {
         return name;
     }
 
-    public Long getPrice() {
+    public BigDecimal getPrice() {
         return price;
     }
