+++ src/main/java/cart/dto/response/ProductResponse.java
@@ -1,14 +1,15 @@
 package cart.dto.response;
 
 import cart.domain.product.Product;
+import java.math.BigDecimal;
 
 public class ProductResponse {
     private Long id;
     private String name;
-    private Long price;
+    private BigDecimal price;
     private String imageUrl;
 
-    private ProductResponse(Long id, String name, Long price, String imageUrl) {
+    private ProductResponse(Long id, String name, BigDecimal price, String imageUrl) {
         this.id = id;
         this.name = name;
         this.price = price;
@@ -32,7 +33,7 @@ public class ProductResponse {
         return name;
     }
 
-    public Long getPrice() {
+    public BigDecimal getPrice() {
         return price;
     }
