+++ src/test/java/cart/domain/order/OrderItemTest.java
@@ -7,6 +7,7 @@ import cart.domain.product.ImageUrl;
 import cart.domain.product.Name;
 import cart.domain.product.Price;
 import cart.domain.product.Product;
+import java.math.BigDecimal;
 import org.junit.jupiter.api.DisplayNameGeneration;
 import org.junit.jupiter.api.DisplayNameGenerator.ReplaceUnderscores;
 import org.junit.jupiter.api.Test;
@@ -19,13 +20,14 @@ public class OrderItemTest {
     void 하나의_주문_목록은_총_금액을_계산할_수_있다() {
         // given
         final Quantity quantity = new Quantity(5);
-        final Product product = new Product(1L, new Name("상품"), new ImageUrl("example.com"), new Price(1000L));
+        final Product product = new Product(1L, new Name("상품"), new ImageUrl("example.com"),
+                new Price(BigDecimal.valueOf(1000)));
         final OrderItem orderItem = new OrderItem(1L, quantity, product);
 
         //when
         final Price totalPrice = orderItem.getTotalPrice();
 
         //then
-        assertThat(totalPrice.price()).isEqualTo(5000);
+        assertThat(totalPrice.price().intValue()).isEqualTo(5000);
     }
 }
